Exclude regions >gamma km away from any case
The main bottleneck for computation is the calculation of the spatial connectivity matrix at each time step, which directly depends on the number of regions in data$distance (and data$population). To reduce the computation time, we remove the regions that have no case reported in a *gamma* radius around them.